finished the main functionalities of project In this project, I wasn't abble to do alone without watch the solution. I never thought that the solution was this. Professor Sahand Gavidel used a little math account to get the wanted effect. The count is this:
document.style.opacity = 1 - window.scrollY / 800 => To erase image as we scroll down;
document.style.backgroundSize = 150(the size setted in CSS) - window.scrollY / 12.

We need to set some value in background-size in CSS to manipulate in the Javascript with the math account showed above.

Now, I'll add some personal touches in this project.